show head of service
Show the head of service. This currently uses a simple link block. We can switch it use one with a profile thumbnail when one is specified once ... is merged.

This also adds a `yarn start:dev` command. It "works" but is a little flaky. When it works, it is delightful though :smiley:
Will look into it and figure out why it sometimes fails. This is an improvement for now :smiley:

fix #17